+++ changelog.md
@@ -1,5 +1,11 @@
 Active Directory \[handlers\] (2019-10-01)
   \[active-directory-user-remove-groups-v3\] (2019-10-01)
-    * Handler added.  Tested on Task 4.0.6
+    * PER-170 Handler added.  Tested on Task 4.0.6
   \[active-directory-user-add-groups-v3\] (2019-10-01)
-    * Handler added.  Tested on Task 4.0.6
+    * PER-170 Handler added.  Tested on Task 4.0.6
+  \[active-directory-computer-remove-groups-v2\] (2019-10-01)
+    * PER-170 Handler added.  Tested on Task 4.0.6
+  \[active-directory-computer-add-groups-v2\] (2019-10-01)
+    * PER-170 Handler added.  Tested on Task 4.0.6
+  \[active-directory-group-create-v3\] (2019-10-01)
+    * PER-170 Handler added.  Tested on Task 4.0.6